ui(refactor): split 'superdesk-groups' module.
