+++ scripts/superdesk-groups/controllers/GroupsSettingsController.js
@@ -1,0 +1,6 @@
+GroupsSettingsController.$inject = ['$scope', 'groups'];
+export function GroupsSettingsController($scope, groups) {
+    groups.initialize().then(function() {
+        $scope.groups = groups.groups;
+    });
+}
