+++ scripts/superdesk-groups/controllers/index.js
@@ -1,0 +1,2 @@
+export { GroupsSettingsController } from './GroupsSettingsController';
+export { GroupsConfigController } from './GroupsConfigController';
