+++ scripts/superdesk-groups/directives/GroupsConfigModal.js
@@ -1,0 +1,6 @@
+export function GroupsConfigModal() {
+    return {
+        require: '^sdGroupsConfig',
+        templateUrl: 'scripts/superdesk-groups/views/groups-config-modal.html'
+    };
+}
