+++ scripts/superdesk-groups/directives/index.js
@@ -1,0 +1,3 @@
+export { GroupeditBasicDirective } from './GroupeditBasicDirective';
+export { GroupeditPeopleDirective } from './GroupeditPeopleDirective';
+export { GroupsConfigModal } from './GroupsConfigModal';
