+++ scripts/superdesk-groups/index.js
@@ -1,5 +1,35 @@
 // styles
 import './styles/groups.less';
 
-// scripts
-import './groups';
+import { GroupsSettingsController, GroupsConfigController } from './controllers';
+import { GroupsFactory } from './services';
+import * as directive from './directives';
+
+export default angular.module('superdesk.groups', [
+    'superdesk.users'
+])
+    .config(['superdeskProvider', function(superdesk) {
+        superdesk
+        .activity('/settings/groups', {
+                label: gettext('Groups'),
+                controller: GroupsSettingsController,
+                templateUrl: 'scripts/superdesk-groups/views/settings.html',
+                category: superdesk.MENU_SETTINGS,
+                priority: -800,
+                beta: true,
+                privileges: {groups: 1}
+            });
+    }])
+    .config(['apiProvider', function(apiProvider) {
+        apiProvider.api('groups', {
+            type: 'http',
+            backend: {
+                rel: 'groups'
+            }
+        });
+    }])
+    .factory('groups', GroupsFactory)
+    .directive('sdGroupeditBasic', directive.GroupeditBasicDirective)
+    .directive('sdGroupeditPeople', directive.GroupeditPeopleDirective)
+    .directive('sdGroupsConfig', () => ({controller: GroupsConfigController}))
+    .directive('sdGroupsConfigModal', directive.GroupsConfigModal);
