+++ scripts/superdesk-groups/services/index.js
@@ -1,0 +1,1 @@
+export { GroupsFactory } from './GroupsFactory';
